(RAZOR-733) Remove unused management ports.
Prior to this commit unused JBoss management ports were listening by default on 9999, 9990, and 9443. These ports, and this management console is unused and deemed an unnecessary security vulnerability.

In this commit the management interface and socket-bindings have been removed - this alone caused an error of missing dependencies. A setting of use-management-endpoint='false' was added to the jmx subsystem remoting-connector to remove the expected dependencies for [service jboss.serverManagement.controller.management.http] and [service jboss.remoting.server.management]